Chore: 기존 코드 폴더링 및 리팩토링
