+++ RPE/Resource/Assets.xcassets/Contents.json
