+++ RPE/Source/Model/RpeData.swift
@@ -1,12 +1,11 @@
 import Foundation
 
 // MARK: - RPE 계산 데이터 
-class RpeDataModel {
+class RpeData {
      
     var rpeArray : [[Double]]
     
     init() {
-        // 초기화
         rpeArray = [
             [1, 0.955, 0.922, 0.892, 0.863, 0.837, 0.811, 0.786, 0.762, 0.739, 0.707, 0.68],
             [0.978, 0.939, 0.907, 0.878, 0.85,  0.824, 0.799, 0.774, 0.751, 0.723, 0.694, 0.667],
@@ -16,7 +15,6 @@ class RpeDataModel {
             [0.907, 0.878, 0.85,  0.824, 0.799, 0.774, 0.751, 0.723, 0.694, 0.667, 0.64, 0.613],
             [0.892, 0.863, 0.837, 0.811, 0.786, 0.762, 0.739, 0.707, 0.68, 0.653, 0.626, 0.599],
             [0.878, 0.85,  0.824, 0.799, 0.774, 0.751, 0.723, 0.694, 0.667, 0.64, 0.613, 0.586]
-        
         ]
     }
