+++ RPE/Source/ViewModel/IsTextViewModel.swift
@@ -3,7 +3,7 @@ import SwiftUI
 // MARK: - Setting화면에서 중량단위 kg <-> lb 변경
 class IsTextViewModel: ObservableObject {
     
-    @Published var isText: Bool /// kg <-> lb
+    @Published var isText: Bool // kg <-> lb
     
     init() {
         self.isText = UserDefaults.standard.bool(forKey: "isText")
